requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
boto==2.40.0
certifi==2016.2.28
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
django-braces==1.8.1
django-jsonview==0.5.0
django-oauth-toolkit==0.10.0
git+https://github.com/EndyKaufman/django-seo-js.git#master
django-storages==1.4.1
django-toolbelt==0.0.1
easy-thumbnails==2.3
gunicorn==19.5.0
lorem-ipsum-generator==0.3
oauthlib==1.0.3
Pillow==3.2.0
psycopg2==2.6.1
requests==2.10.0
shortuuid==0.4.3
singledispatch==3.4.0.3
six==1.10.0
static3==0.7.0
tornado==4.3
transliterate==1.7.6
whitenoise==3.0
solid-i18n==1.3.0
django-modeltranslation==0.11
